+++ packages/iov-ledger-bns/src/app.spec.ts
@@ -108,7 +108,7 @@ describe("Sign with ledger app", () => {
     const tx: SendTx = {
       kind: TransactionKind.Send,
       chainId: "test-bns-ledger" as ChainId,
-      recipient: Encoding.fromHex("0123456789012345678901234567890123456789") as RecipientId,
+      recipient: "0123456789012345678901234567890123456789" as RecipientId,
       amount: {
         // 1234.789 LGR
         whole: 1234,
@@ -153,7 +153,7 @@ describe("Sign with ledger app", () => {
     const tx: SendTx = {
       kind: TransactionKind.Send,
       chainId: "test-ledger-paths" as ChainId,
-      recipient: Encoding.fromHex("1234ABCD0000AA0000FFFF0000AA00001234ABCD") as RecipientId,
+      recipient: "1234ABCD0000AA0000FFFF0000AA00001234ABCD" as RecipientId,
       amount: {
         // 77.01001 PATH
         whole: 77,