+++ packages/iov-lisk/src/derivation.spec.ts
@@ -2,7 +2,7 @@ import { Encoding } from "@iov/encoding";
 
 import { passphraseToKeypair, pubkeyToAddress } from "./derivation";
 
-const { fromHex, toAscii } = Encoding;
+const { fromHex } = Encoding;
 
 describe("Derivation", () => {
   describe("passphraseToKeypair", () => {
@@ -23,13 +23,13 @@ describe("Derivation", () => {
     it("works for address in signed int64 range", () => {
       // https://testnet-explorer.lisk.io/address/6076671634347365051L
       const pubkey = fromHex("f4852b270f76dc8b49bfa88de5906e81d3b001d23852f0e74ba60cac7180a184");
-      expect(pubkeyToAddress(pubkey)).toEqual(toAscii("6076671634347365051L"));
+      expect(pubkeyToAddress(pubkey)).toEqual("6076671634347365051L");
     });
 
     it("works for address outside of signed int64 range", () => {
       // https://testnet-explorer.lisk.io/address/10176009299933723198L
       const pubkey = fromHex("06ad4341a609af2de837e1156f81849b05bf3c280940a9f45db76d09a3a3f2fa");
-      expect(pubkeyToAddress(pubkey)).toEqual(toAscii("10176009299933723198L"));
+      expect(pubkeyToAddress(pubkey)).toEqual("10176009299933723198L");
     });
   });
 });